refactor: split DocumentsPage into focused sub-components
Extract 6 components and 3 hooks from the 1194-line index.tsx:
- RecoveryBanner: recovered upload sessions banner
- UploadZone: drag-drop upload area with compact/full modes
- DocumentSearchBar: search input, type filter, view toggle
- BatchActionsBar: multi-select batch move/delete bar
- DocumentListContainer: list/grid rendering, pagination, empty states
- useDocumentPreview: preview modal state and navigation
- useMultiSelect: multi-select state and batch operations
- useEditNavigation: edit modal navigation and delete-from-edit

Index.tsx reduced from 1194 to 415 lines. No behavior changes.
