+++ lexwebapp/src/pages/DocumentsPage/index.tsx
@@ -1,106 +1,36 @@
-import React, { useState, useRef, useCallback, useEffect, useMemo } from 'react';
+import { useState, useRef, useCallback, useEffect, useMemo } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
-import { motion, AnimatePresence } from 'framer-motion';
-import {
-  Upload,
-  FolderUp,
-  FileText,
-  Search,
-  X,
-  CheckCircle,
-  AlertCircle,
-  Loader2,
-  LayoutGrid,
-  List,
-  Trash2,
-  FolderInput,
-  SearchX,
-} from 'lucide-react';
-import { mcpService } from '../../services';
 import { useShallow } from 'zustand/react/shallow';
 import { useUploadStore } from '../../stores/uploadStore';
 import { showToast } from '../../utils/toast';
-import { api } from '../../utils/api-client';
 import { FolderNavigator } from './FolderNavigator';
 import { UploadQueuePanel } from './UploadQueuePanel';
-import { DocumentTable } from './DocumentTable';
-import { DocumentGrid } from './DocumentGrid';
 import { DocumentViewerModal } from '../../components/DocumentViewerModal';
 import { ClassificationPanel } from './ClassificationPanel';
 import { DeleteConfirmModal } from './DeleteConfirmModal';
 import { EditDocumentModal } from './EditDocumentModal';
 import { MoveDocumentModal } from './MoveDocumentModal';
-import type { VaultDocument, DocType, ViewMode, SortField, SortOrder } from './types';
-import { isPreviewableBinary } from './types';
-import { processEmlContent } from '../../utils/eml-parser';
+import { RecoveryBanner } from './RecoveryBanner';
+import { UploadZone } from './UploadZone';
+import { DocumentSearchBar } from './DocumentSearchBar';
+import { BatchActionsBar } from './BatchActionsBar';
+import { DocumentListContainer } from './DocumentListContainer';
+import type { DocType, ViewMode, SortField, SortOrder } from './types';
 import { useUndoStore } from '../../stores/undoStore';
 import { useUndoKeyboard } from '../../hooks/useUndoKeyboard';
 import { useDocumentData } from './useDocumentData';
 import { useDocumentActions } from './useDocumentActions';
-import { DOC_TYPE_LABELS } from './constants';
-
-const ACCEPTED_TYPES =
-  '.pdf,.docx,.doc,.html,.htm,.txt,.rtf,.jpg,.jpeg,.png,.bmp,.gif,.xlsx,.xls,.csv,.mp4,.mov,.avi,.mkv,.webm,.eml,.zip,.gz,.tgz,.tar';
+import { useDocumentPreview } from './useDocumentPreview';
+import { useMultiSelect } from './useMultiSelect';
+import { useEditNavigation } from './useEditNavigation';
 
 const VIEW_MODE_KEY = 'documents-view-mode';
-
-function guessMimeType(file: File): string {
-  if (file.type) return file.type;
-  const ext = file.name.split('.').pop()?.toLowerCase();
-  const map: Record<string, string> = {
-    pdf: 'application/pdf',
-    docx: 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-    doc: 'application/msword',
-    html: 'text/html',
-    htm: 'text/html',
-    txt: 'text/plain',
-    rtf: 'application/rtf',
-    jpg: 'image/jpeg',
-    jpeg: 'image/jpeg',
-    png: 'image/png',
-    bmp: 'image/bmp',
-    gif: 'image/gif',
-    xlsx: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-    xls: 'application/vnd.ms-excel',
-    csv: 'text/csv',
-    mp4: 'video/mp4',
-    mov: 'video/quicktime',
-    avi: 'video/x-msvideo',
-    mkv: 'video/x-matroska',
-    webm: 'video/webm',
-    eml: 'message/rfc822',
-    zip: 'application/zip',
-    gz: 'application/gzip',
-    tgz: 'application/x-tgz',
-    tar: 'application/x-tar',
-  };
-  return map[ext || ''] || 'application/octet-stream';
-}
-
-function SkeletonRows() {
-  return (
-    <div className="bg-white rounded-2xl border border-claude-border shadow-sm overflow-hidden">
-      <div className="divide-y divide-claude-border/30">
-        {Array.from({ length: 5 }).map((_, i) => (
-          <div key={i} className="flex items-center gap-4 px-5 py-4 animate-pulse">
-            <div className="h-4 w-4 bg-claude-border/30 rounded" />
-            <div className="h-4 bg-claude-border/30 rounded w-1/3" />
-            <div className="h-4 bg-claude-border/30 rounded w-16" />
-            <div className="h-4 bg-claude-border/30 rounded w-24" />
-            <div className="h-4 bg-claude-border/30 rounded w-14 hidden md:block" />
-            <div className="h-4 bg-claude-border/30 rounded w-20 hidden md:block" />
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
 
 export function DocumentsPage() {
   const navigate = useNavigate();
   const params = useParams();
 
-  // Derive folder path from URL: /documents/folders/a/b → "a/b/"
+  // Derive folder path from URL: /documents/folders/a/b -> "a/b/"
   const currentFolderPath = useMemo(() => {
     const wildcard = params['*'] || '';
     if (!wildcard) return '';
@@ -118,25 +48,6 @@ export function DocumentsPage() {
   const [sortBy, setSortBy] = useState<SortField>('uploadedAt');
   const [sortOrder, setSortOrder] = useState<SortOrder>('desc');
 
-  // Multi-select state
-  const [selectedIds, setSelectedIds] = useState<Set<string>>(new Set());
-
-  // Preview state
-  const [previewDoc, setPreviewDoc] = useState<{
-    type: 'document';
-    title: string;
-    subtitle?: string;
-    badge?: string;
-    content: string;
-    previewUrl?: string;
-    mimeType?: string;
-    ocrText?: string;
-    documentId?: string;
-  } | null>(null);
-  const [previewOpen, setPreviewOpen] = useState(false);
-  const [previewLoading, setPreviewLoading] = useState(false);
-  const [previewIndex, setPreviewIndex] = useState<number>(-1);
-
   // Compact upload zone
   const [uploadZoneExpanded, setUploadZoneExpanded] = useState(false);
 
@@ -150,11 +61,6 @@ export function DocumentsPage() {
   useEffect(() => {
     setOffset(0);
   }, [filterType, currentFolderPath, searchQuery]);
-
-  // Clear selection when documents change
-  useEffect(() => {
-    setSelectedIds(new Set());
-  }, [currentFolderPath, filterType, searchQuery, offset]);
 
   // Data fetching hook
   const {
@@ -196,7 +102,67 @@ export function DocumentsPage() {
     handleSaveOcrText, pushAction, undoAction,
   } = actions;
 
-  // Upload state from Zustand store — shallow selector for state values
+  // Preview hook
+  const preview = useDocumentPreview({
+    documents,
+    setDocuments,
+    setTotalDocs,
+    offset,
+    setOffset,
+    totalDocs,
+    PAGE_SIZE,
+    sortBy,
+    sortOrder,
+    filterType,
+    currentFolderPath,
+    searchQuery,
+    pushAction,
+    undoAction,
+    loadDocuments,
+    loadFolders,
+  });
+
+  // Multi-select hook
+  const {
+    selectedIds,
+    handleToggleSelect,
+    handleToggleAll,
+    handleBatchDelete,
+    handleBatchMove,
+    clearSelection,
+  } = useMultiSelect({
+    documents,
+    currentFolderPath,
+    filterType,
+    searchQuery,
+    offset,
+    pushAction,
+    loadDocuments,
+    loadFolders,
+    handleMoveOpen,
+  });
+
+  // Edit navigation hook
+  const {
+    editIndex,
+    handleEditPrevious,
+    handleEditNext,
+    handleEditDelete,
+  } = useEditNavigation({
+    documents,
+    editTarget,
+    handleEditOpen,
+    setEditTarget,
+    setDocuments,
+    setTotalDocs,
+    pushAction,
+    undoAction,
+    loadDocuments,
+    loadFolders,
+    currentFolderPath,
+  });
+
+  // Upload state from Zustand store
   const { items: uploadItems, isUploading, completedFiles, recoveredSessions } = useUploadStore(
     useShallow(s => ({
       items: s.items,
@@ -215,7 +181,6 @@ export function DocumentsPage() {
   const setOnActionExecuted = useUndoStore((s) => s.setOnActionExecuted);
   useUndoKeyboard();
 
-  // Register reload callback for undo/redo
   useEffect(() => {
     setOnActionExecuted(() => {
       loadDocuments();
@@ -234,23 +199,16 @@ export function DocumentsPage() {
   const dropZoneRef = useRef<HTMLDivElement>(null);
 
   // Check for stuck upload sessions on mount
-  useEffect(() => {
-    recoverSessions();
-  }, []);
+  useEffect(() => { recoverSessions(); }, []);
 
   // Reload docs when a recovered session completes
   useEffect(() => {
-    const hasNewlyCompleted = recoveredSessions.some((s) => s.status === 'completed');
-    if (hasNewlyCompleted) {
-      loadDocuments();
-    }
+    if (recoveredSessions.some((s) => s.status === 'completed')) loadDocuments();
   }, [recoveredSessions]);
 
   // Show upload panel when items are added
   useEffect(() => {
-    if (uploadItems.length > 0) {
-      setShowUploadPanel(true);
-    }
+    if (uploadItems.length > 0) setShowUploadPanel(true);
   }, [uploadItems.length]);
 
   // Reload docs and folders when uploads complete
@@ -264,97 +222,16 @@ export function DocumentsPage() {
 
   // Navigation helpers
   const navigateToFolder = useCallback((folderPath: string) => {
-    if (!folderPath) {
-      navigate('/documents');
-    } else {
-      const cleanPath = folderPath.replace(/\/+$/, '');
-      navigate(`/documents/folders/${cleanPath}`);
-    }
+    if (!folderPath) navigate('/documents');
+    else navigate(`/documents/folders/${folderPath.replace(/\/+$/, '')}`);
   }, [navigate]);
 
-  // File selection handlers
-  const handleFilesSelected = useCallback(
-    (files: FileList | File[]) => {
-      const newItems = Array.from(files)
-        .filter((f) => f.size > 0)
-        .map((file) => ({
-          file,
-          mimeType: guessMimeType(file),
-          relativePath: (file as any).webkitRelativePath || currentFolderPath || '',
-          docType: defaultDocType,
-        }));
-
-      if (newItems.length === 0) return;
-      addFiles(newItems);
-      setShowUploadPanel(true);
-    },
-    [defaultDocType, addFiles, currentFolderPath]
-  );
-
-  const handleFileSelect = () => fileInputRef.current?.click();
-  const handleFolderSelect = () => folderInputRef.current?.click();
-
-  const handleFileInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (e.target.files?.length) {
-      handleFilesSelected(e.target.files);
-      e.target.value = '';
-    }
-  };
-
-  // Drag and drop
-  const handleDragOver = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragOver(true);
-    setUploadZoneExpanded(true);
-  };
-  const handleDragLeave = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (dropZoneRef.current && !dropZoneRef.current.contains(e.relatedTarget as Node)) {
-      setIsDragOver(false);
-    }
-  };
-  const handleDrop = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragOver(false);
-
-    const items = e.dataTransfer.items;
-    if (items) {
-      const files: File[] = [];
-      const entries: any[] = [];
-      for (let i = 0; i < items.length; i++) {
-        const entry = (items[i] as any).webkitGetAsEntry?.();
-        if (entry) {
-          entries.push(entry);
-        } else if (items[i].kind === 'file') {
-          const f = items[i].getAsFile();
-          if (f) files.push(f);
-        }
-      }
-
-      if (entries.length > 0) {
-        readEntries(entries).then((allFiles) => {
-          handleFilesSelected(allFiles);
-        });
-      } else if (files.length > 0) {
-        handleFilesSelected(files);
-      }
-    } else if (e.dataTransfer.files.length) {
-      handleFilesSelected(e.dataTransfer.files);
-    }
-  };
-
-  // Upload actions
   const handleStartUpload = () => {
     startUpload();
     showToast.success(`Завантаження ${uploadItems.filter((i) => i.status === 'queued').length} файлів розпочато`);
   };
 
-  const handleSearch = async () => {
-    loadDocuments();
-  };
+  const handleSearch = () => { loadDocuments(); };
 
   const handleClearSearch = () => {
     setSearchQuery('');
@@ -362,490 +239,50 @@ export function DocumentsPage() {
     loadDocuments();
   };
 
-  // Sort handler
   const handleSort = (field: SortField) => {
-    if (field === sortBy) {
-      setSortOrder((prev) => (prev === 'asc' ? 'desc' : 'asc'));
-    } else {
-      setSortBy(field);
-      setSortOrder('desc');
-    }
+    if (field === sortBy) setSortOrder((prev) => (prev === 'asc' ? 'desc' : 'asc'));
+    else { setSortBy(field); setSortOrder('desc'); }
   };
-
-  // Multi-select handlers
-  const handleToggleSelect = useCallback((id: string) => {
-    setSelectedIds((prev) => {
-      const next = new Set(prev);
-      if (next.has(id)) next.delete(id);
-      else next.add(id);
-      return next;
-    });
-  }, []);
-
-  const handleToggleAll = useCallback(() => {
-    setSelectedIds((prev) => {
-      const allSelected = documents.every((d) => prev.has(d.id));
-      if (allSelected) return new Set();
-      return new Set(documents.map((d) => d.id));
-    });
-  }, [documents]);
-
-  const handleBatchDelete = useCallback(async () => {
-    if (selectedIds.size === 0) return;
-    const count = selectedIds.size;
-    const ids = Array.from(selectedIds);
-    try {
-      await Promise.all(ids.map((id) => api.documents.delete(id)));
-      ids.forEach((id) => {
-        const doc = documents.find((d) => d.id === id);
-        if (doc) pushAction({ type: 'delete', documentId: id, documentTitle: doc.title });
-      });
-      showToast.success(`${count} документів видалено`);
-      setSelectedIds(new Set());
-      loadDocuments();
-      loadFolders(currentFolderPath);
-    } catch (err) {
-      console.error('Batch delete failed:', err);
-      showToast.error('Не вдалося видалити деякі документи');
-    }
-  }, [selectedIds, documents, pushAction, loadDocuments, loadFolders, currentFolderPath]);
-
-  const handleBatchMove = useCallback(() => {
-    if (selectedIds.size === 0) return;
-    // Use the first selected document as move target to open modal
-    const firstDoc = documents.find((d) => selectedIds.has(d.id));
-    if (firstDoc) handleMoveOpen(firstDoc);
-  }, [selectedIds, documents, handleMoveOpen]);
-
-  // Document preview
-  const handleDocumentClick = async (doc: VaultDocument, index?: number) => {
-    if (index != null) setPreviewIndex(index);
-    else setPreviewIndex(documents.findIndex((d) => d.id === doc.id));
-    const mimeType = doc.metadata?.mimeType || doc.mime_type;
-
-    // For binary previewable files (images, PDFs, videos), fetch presigned URL
-    if (isPreviewableBinary(mimeType)) {
-      setPreviewLoading(true);
-      setPreviewOpen(true);
-      try {
-        const [previewResp, docResp] = await Promise.all([
-          api.documents.getPreviewUrl(doc.id),
-          api.documents.getById(doc.id).catch(() => null),
-        ]);
-
-        const { previewUrl, mimeType: serverMime } = previewResp.data;
-        const badge = DOC_TYPE_LABELS[doc.type] || doc.type;
-        const ocrText = docResp?.data?.full_text ?? undefined;
-        const effectiveMime = serverMime || mimeType;
-        const isImagePreview = effectiveMime?.startsWith('image/');
-
-        if (previewUrl) {
-          setPreviewDoc({
-            type: 'document',
-            title: doc.title,
-            subtitle: doc.metadata?.uploadedAt ? `Завантажено: ${new Date(doc.metadata.uploadedAt).toLocaleDateString('uk-UA')}` : undefined,
-            badge,
-            content: ocrText || '',
-            previewUrl,
-            mimeType: effectiveMime,
-            ...(isImagePreview ? { ocrText: ocrText || '', documentId: doc.id } : {}),
-          });
-        } else {
-          try {
-            const result = await mcpService.callTool('get_document', { documentId: doc.id });
-            const parsed = result?.result?.content?.[0]?.text
-              ? JSON.parse(result.result.content[0].text)
-              : result?.result || result;
-            const rawContent = parsed.content || parsed.text || parsed.sections?.map((s: any) => s.content).join('\n\n') || 'Вміст недоступний';
-            const content = processEmlContent(rawContent);
-            setPreviewDoc({
-              type: 'document',
-              title: doc.title,
-              subtitle: doc.metadata?.uploadedAt ? `Завантажено: ${new Date(doc.metadata.uploadedAt).toLocaleDateString('uk-UA')}` : undefined,
-              badge,
-              content,
-            });
-          } catch {
-            setPreviewDoc({
-              type: 'document',
-              title: doc.title,
-              subtitle: doc.metadata?.uploadedAt ? `Завантажено: ${new Date(doc.metadata.uploadedAt).toLocaleDateString('uk-UA')}` : undefined,
-              badge,
-              content: 'Попередній перегляд недоступний для цього файлу.',
-            });
-          }
-        }
-      } catch (err) {
-        console.error('Failed to fetch preview URL:', err);
-        showToast.error('Не вдалося завантажити попередній перегляд');
-        setPreviewOpen(false);
-      } finally {
-        setPreviewLoading(false);
-      }
-      return;
-    }
-
-    // For text/document files, load full text via MCP tool
-    setPreviewLoading(true);
-    setPreviewOpen(true);
-    try {
-      const result = await mcpService.callTool('get_document', { documentId: doc.id });
-      const parsed = result?.result?.content?.[0]?.text
-        ? JSON.parse(result.result.content[0].text)
-        : result?.result || result;
-
-      const rawContent = parsed.content || parsed.text || parsed.sections?.map((s: any) => s.content).join('\n\n') || 'Вміст недоступний';
-      const content = processEmlContent(rawContent);
-      const badge = DOC_TYPE_LABELS[doc.type] || doc.type;
-
-      setPreviewDoc({
-        type: 'document',
-        title: doc.title,
-        subtitle: doc.metadata?.uploadedAt ? `Завантажено: ${new Date(doc.metadata.uploadedAt).toLocaleDateString('uk-UA')}` : undefined,
-        badge,
-        content,
-      });
-    } catch (err) {
-      console.error('Failed to fetch document:', err);
-      showToast.error('Не вдалося завантажити документ');
-      setPreviewOpen(false);
-    } finally {
-      setPreviewLoading(false);
-    }
-  };
-
-  // Preview navigation handlers (with auto-pagination)
-  const handlePreviewPrevious = useCallback(async () => {
-    if (previewIndex > 0) {
-      const prevDoc = documents[previewIndex - 1];
-      if (prevDoc) handleDocumentClick(prevDoc, previewIndex - 1);
-    } else if (offset > 0) {
-      const newOffset = Math.max(0, offset - PAGE_SIZE);
-      setOffset(newOffset);
-      setPreviewLoading(true);
-      try {
-        const params: Record<string, any> = { limit: PAGE_SIZE, offset: newOffset, sortBy, sortOrder };
-        if (filterType) params.type = filterType;
-        if (currentFolderPath) params.folderPath = currentFolderPath;
-        if (searchQuery.trim()) params.query = searchQuery.trim();
-        const result = await mcpService.callTool('list_documents', params);
-        const parsed = result?.result?.content?.[0]?.text ? JSON.parse(result.result.content[0].text) : result?.result || result;
-        const newDocs: VaultDocument[] = parsed.documents || [];
-        setDocuments(newDocs);
-        setTotalDocs(parsed.total || 0);
-        if (newDocs.length > 0) {
-          const lastDoc = newDocs[newDocs.length - 1];
-          handleDocumentClick(lastDoc, newDocs.length - 1);
-        }
-      } catch (err) {
-        console.error('Failed to load previous page:', err);
-        setPreviewLoading(false);
-      }
-    }
-  }, [previewIndex, documents, offset, sortBy, sortOrder, filterType, currentFolderPath, searchQuery]);
-
-  const handlePreviewNext = useCallback(async () => {
-    if (previewIndex < documents.length - 1) {
-      const nextDoc = documents[previewIndex + 1];
-      if (nextDoc) handleDocumentClick(nextDoc, previewIndex + 1);
-    } else if (offset + PAGE_SIZE < totalDocs) {
-      const newOffset = offset + PAGE_SIZE;
-      setOffset(newOffset);
-      setPreviewLoading(true);
-      try {
-        const params: Record<string, any> = { limit: PAGE_SIZE, offset: newOffset, sortBy, sortOrder };
-        if (filterType) params.type = filterType;
-        if (currentFolderPath) params.folderPath = currentFolderPath;
-        if (searchQuery.trim()) params.query = searchQuery.trim();
-        const result = await mcpService.callTool('list_documents', params);
-        const parsed = result?.result?.content?.[0]?.text ? JSON.parse(result.result.content[0].text) : result?.result || result;
-        const newDocs: VaultDocument[] = parsed.documents || [];
-        setDocuments(newDocs);
-        setTotalDocs(parsed.total || 0);
-        if (newDocs.length > 0) {
-          handleDocumentClick(newDocs[0], 0);
-        }
-      } catch (err) {
-        console.error('Failed to load next page:', err);
-        setPreviewLoading(false);
-      }
-    }
-  }, [previewIndex, documents, offset, totalDocs, sortBy, sortOrder, filterType, currentFolderPath, searchQuery]);
-
-  // Delete from preview modal — with double-click confirmation
-  const [previewDeletePending, setPreviewDeletePending] = useState(false);
-  const handlePreviewDelete = useCallback(async () => {
-    if (previewIndex < 0 || !documents[previewIndex]) return;
-
-    // First press = arm, second press = delete
-    if (!previewDeletePending) {
-      setPreviewDeletePending(true);
-      showToast.success('Натисніть Delete ще раз для підтвердження');
-      setTimeout(() => setPreviewDeletePending(false), 3000);
-      return;
-    }
-
-    setPreviewDeletePending(false);
-    const doc = documents[previewIndex];
-    try {
-      await api.documents.delete(doc.id);
-      pushAction({ type: 'delete', documentId: doc.id, documentTitle: doc.title });
-      showToast.undoable(`«${doc.title}» видалено`, () => {
-        undoAction().then(() => { loadDocuments(); loadFolders(currentFolderPath); });
-      });
-      const newDocs = documents.filter((_, i) => i !== previewIndex);
-      setDocuments(newDocs);
-      setTotalDocs((prev) => Math.max(0, prev - 1));
-      if (newDocs.length === 0) {
-        setPreviewOpen(false);
-        setPreviewDoc(null);
-        setPreviewIndex(-1);
-      } else if (previewIndex < newDocs.length) {
-        handleDocumentClick(newDocs[previewIndex], previewIndex);
-      } else {
-        handleDocumentClick(newDocs[newDocs.length - 1], newDocs.length - 1);
-      }
-    } catch (err) {
-      console.error('Failed to delete document:', err);
-      showToast.error('Не вдалося видалити документ');
-    }
-  }, [previewIndex, documents, previewDeletePending]);
-
-  // Edit navigation handlers
-  const editIndex = editTarget ? documents.findIndex((d) => d.id === editTarget.id) : -1;
-
-  const handleEditPrevious = useCallback(() => {
-    if (editIndex > 0) handleEditOpen(documents[editIndex - 1]);
-  }, [editIndex, documents]);
-
-  const handleEditNext = useCallback(() => {
-    if (editIndex < documents.length - 1) handleEditOpen(documents[editIndex + 1]);
-  }, [editIndex, documents]);
-
-  // Delete from edit modal — with double-click confirmation
-  const handleEditDelete = useCallback(async () => {
-    if (!editTarget) return;
-    const idx = documents.findIndex((d) => d.id === editTarget.id);
-    try {
-      await api.documents.delete(editTarget.id);
-      pushAction({ type: 'delete', documentId: editTarget.id, documentTitle: editTarget.title });
-      showToast.undoable(`«${editTarget.title}» видалено`, () => {
-        undoAction().then(() => { loadDocuments(); loadFolders(currentFolderPath); });
-      });
-      const newDocs = documents.filter((d) => d.id !== editTarget.id);
-      setDocuments(newDocs);
-      setTotalDocs((prev) => Math.max(0, prev - 1));
-      if (newDocs.length === 0) {
-        setEditTarget(null);
-      } else if (idx < newDocs.length) {
-        handleEditOpen(newDocs[idx]);
-      } else {
-        handleEditOpen(newDocs[newDocs.length - 1]);
-      }
-    } catch (err) {
-      console.error('Failed to delete document:', err);
-      showToast.error('Не вдалося видалити документ');
-    }
-  }, [editTarget, documents]);
 
   // Keyboard shortcut for upload (Ctrl+U)
   useEffect(() => {
     const handleKeydown = (e: KeyboardEvent) => {
       if ((e.ctrlKey || e.metaKey) && e.key === 'u') {
         e.preventDefault();
-        handleFileSelect();
+        fileInputRef.current?.click();
       }
     };
     window.addEventListener('keydown', handleKeydown);
     return () => window.removeEventListener('keydown', handleKeydown);
   }, []);
 
-  // Pagination
-  const hasMore = offset + PAGE_SIZE < totalDocs;
-  const hasPrev = offset > 0;
-  const isSearchActive = searchQuery.trim().length > 0;
   const hasDocuments = documents.length > 0;
   const showCompactUpload = hasDocuments && !uploadZoneExpanded;
 
   return (
-    <div
-      className="flex-1 flex flex-col h-full overflow-hidden"
-      onDragOver={handleDragOver}
-      onDragLeave={handleDragLeave}
-      onDrop={handleDrop}
-    >
-      {/* Hidden file inputs */}
-      <input
-        ref={fileInputRef}
-        type="file"
-        multiple
-        accept={ACCEPTED_TYPES}
-        className="hidden"
-        onChange={handleFileInputChange}
-      />
-      <input
-        ref={folderInputRef}
-        type="file"
-        multiple
-        accept={ACCEPTED_TYPES}
-        className="hidden"
-        onChange={handleFileInputChange}
-        {...({ webkitdirectory: '', directory: '' } as any)}
-      />
-
+    <div className="flex-1 flex flex-col h-full overflow-hidden">
       <div className="flex-1 overflow-y-auto">
         <div className="max-w-6xl mx-auto px-4 sm:px-6 py-6">
-          {/* Recovery Banner */}
-          <AnimatePresence>
-            {recoveredSessions.length > 0 && (
-              <motion.div
-                initial={{ opacity: 0, height: 0 }}
-                animate={{ opacity: 1, height: 'auto' }}
-                exit={{ opacity: 0, height: 0 }}
-                className="mb-4 bg-amber-50 border border-amber-200 rounded-xl p-4"
-              >
-                <div className="flex items-center justify-between mb-2">
-                  <h4 className="text-sm font-semibold text-amber-800 font-sans">
-                    Відновлення завантажень
-                  </h4>
-                  {recoveredSessions.every((s) => s.status !== 'recovering') && (
-                    <button
-                      onClick={clearRecoveredSessions}
-                      className="text-xs text-amber-600 hover:text-amber-800 transition-colors font-sans"
-                    >
-                      Закрити
-                    </button>
-                  )}
-                </div>
-                <div className="space-y-1.5">
-                  {recoveredSessions.map((s) => (
-                    <div key={s.uploadId} className="flex items-center gap-2 text-xs font-sans">
-                      {s.status === 'recovering' ? (
-                        <Loader2 size={12} className="text-amber-500 animate-spin flex-shrink-0" />
-                      ) : s.status === 'completed' ? (
-                        <CheckCircle size={12} className="text-green-500 flex-shrink-0" />
-                      ) : (
-                        <AlertCircle size={12} className="text-red-500 flex-shrink-0" />
-                      )}
-                      <span className="text-amber-900 truncate flex-1">{s.fileName}</span>
-                      <span className={`flex-shrink-0 ${
-                        s.status === 'recovering' ? 'text-amber-600' :
-                        s.status === 'completed' ? 'text-green-600' : 'text-red-600'
-                      }`}>
-                        {s.status === 'recovering' ? 'Обробка...' :
-                         s.status === 'completed' ? 'Готово' : s.error || 'Помилка'}
-                      </span>
-                      {s.status !== 'recovering' && (
-                        <button
-                          onClick={() => dismissRecoveredSession(s.uploadId)}
-                          className="text-amber-400 hover:text-amber-700 transition-colors"
-                        >
-                          <X size={12} />
-                        </button>
-                      )}
-                    </div>
-                  ))}
-                </div>
-              </motion.div>
-            )}
-          </AnimatePresence>
+          <RecoveryBanner
+            recoveredSessions={recoveredSessions}
+            onDismiss={dismissRecoveredSession}
+            onClearAll={clearRecoveredSessions}
+          />
 
-          {/* Upload Zone — compact when documents exist, full when empty or expanded */}
-          <div ref={dropZoneRef}>
-            {showCompactUpload ? (
-              /* Compact upload toolbar */
-              <div className={`mb-4 transition-all ${isDragOver ? 'ring-2 ring-claude-accent ring-offset-2 rounded-xl' : ''}`}>
-                <div className="flex items-center gap-2">
-                  <button
-                    onClick={handleFileSelect}
-                    className="inline-flex items-center gap-2 px-4 py-2 bg-claude-text text-white rounded-xl text-sm font-medium hover:bg-claude-text/90 transition-all active:scale-[0.98] shadow-sm font-sans"
-                  >
-                    <Upload size={14} strokeWidth={2} />
-                    Завантажити
-                  </button>
-                  <button
-                    onClick={handleFolderSelect}
-                    className="inline-flex items-center gap-1.5 px-3 py-2 bg-white border border-claude-border text-claude-subtext rounded-xl text-sm font-medium hover:bg-claude-bg transition-all active:scale-[0.98] font-sans"
-                  >
-                    <FolderUp size={14} strokeWidth={2} />
-                    Папку
-                  </button>
-                  <span className="text-xs text-claude-subtext/40 font-sans ml-1 hidden sm:inline">
-                    або перетягніть файли · Ctrl+U
-                  </span>
-                </div>
+          <UploadZone
+            isDragOver={isDragOver}
+            setIsDragOver={setIsDragOver}
+            showCompactUpload={showCompactUpload}
+            setUploadZoneExpanded={setUploadZoneExpanded}
+            currentFolderPath={currentFolderPath}
+            defaultDocType={defaultDocType}
+            addFiles={addFiles}
+            setShowUploadPanel={setShowUploadPanel}
+            dropZoneRef={dropZoneRef}
+            fileInputRef={fileInputRef}
+            folderInputRef={folderInputRef}
+          />
 
-                {/* Drag overlay for compact mode */}
-                <AnimatePresence>
-                  {isDragOver && (
-                    <motion.div
-                      initial={{ opacity: 0, height: 0 }}
-                      animate={{ opacity: 1, height: 'auto' }}
-                      exit={{ opacity: 0, height: 0 }}
-                      className="mt-2 flex items-center justify-center bg-claude-accent/10 rounded-xl py-6"
-                    >
-                      <div className="text-claude-accent font-semibold text-lg font-sans">
-                        Відпустіть для завантаження
-                      </div>
-                    </motion.div>
-                  )}
-                </AnimatePresence>
-              </div>
-            ) : (
-              /* Full upload zone */
-              <div
-                className={`relative rounded-2xl border-2 border-dashed p-8 mb-6 transition-all duration-300 ${
-                  isDragOver
-                    ? 'border-claude-accent bg-claude-accent/5 scale-[1.01]'
-                    : 'border-claude-border hover:border-claude-subtext/40 bg-white'
-                }`}
-                role="button"
-                aria-label="Зона завантаження файлів"
-                tabIndex={0}
-                onKeyDown={(e) => { if (e.key === 'Enter' || e.key === ' ') handleFileSelect(); }}
-              >
-                <div className="text-center">
-                  <div className="flex justify-center gap-3 mb-4">
-                    <button
-                      onClick={handleFileSelect}
-                      className="inline-flex items-center gap-2 px-5 py-2.5 bg-claude-text text-white rounded-xl text-sm font-medium hover:bg-claude-text/90 transition-all active:scale-[0.98] shadow-sm"
-                    >
-                      <Upload size={16} strokeWidth={2} />
-                      Завантажити файли
-                    </button>
-                    <button
-                      onClick={handleFolderSelect}
-                      className="inline-flex items-center gap-2 px-5 py-2.5 bg-white border border-claude-border text-claude-text rounded-xl text-sm font-medium hover:bg-claude-bg transition-all active:scale-[0.98] shadow-sm"
-                    >
-                      <FolderUp size={16} strokeWidth={2} />
-                      Завантажити папку
-                    </button>
-                  </div>
-                  <p className="text-sm text-claude-subtext/70 font-sans">
-                    Перетягніть файли або папку сюди &middot; PDF, DOCX, HTML, TXT, зображення, відео &middot; до 2 ГБ &middot; Ctrl+U
-                  </p>
-                </div>
-
-                {/* Drag overlay */}
-                <AnimatePresence>
-                  {isDragOver && (
-                    <motion.div
-                      initial={{ opacity: 0 }}
-                      animate={{ opacity: 1 }}
-                      exit={{ opacity: 0 }}
-                      className="absolute inset-0 flex items-center justify-center bg-claude-accent/10 rounded-2xl z-10"
-                    >
-                      <div className="text-claude-accent font-semibold text-lg font-sans">
-                        Відпустіть для завантаження
-                      </div>
-                    </motion.div>
-                  )}
-                </AnimatePresence>
-              </div>
-            )}
-          </div>
-
-          {/* Upload Queue Panel */}
           <UploadQueuePanel
             showUploadPanel={showUploadPanel}
             setShowUploadPanel={setShowUploadPanel}
@@ -854,122 +291,29 @@ export function DocumentsPage() {
             onStartUpload={handleStartUpload}
           />
 
-          {/* Classification Panel */}
           <ClassificationPanel
             stats={docStats}
-            onComplete={() => {
-              loadDocuments();
-              loadStats();
-            }}
+            onComplete={() => { loadDocuments(); loadStats(); }}
           />
 
-          {/* Search and Filters */}
-          <div className="flex items-center gap-3 mb-5">
-            <div className="flex-1 relative">
-              <Search
-                size={16}
-                className="absolute left-3.5 top-1/2 -translate-y-1/2 text-claude-subtext/40"
-              />
-              <input
-                type="text"
-                placeholder="Пошук за назвою або змістом..."
-                value={searchQuery}
-                onChange={(e) => setSearchQuery(e.target.value)}
-                onKeyDown={(e) => e.key === 'Enter' && handleSearch()}
-                className="w-full pl-10 pr-9 py-2.5 bg-white border border-claude-border rounded-xl text-sm text-claude-text placeholder:text-claude-subtext/40 focus:outline-none focus:border-claude-subtext/40 transition-colors font-sans"
-              />
-              {searchQuery && (
-                <button
-                  onClick={handleClearSearch}
-                  className="absolute right-3 top-1/2 -translate-y-1/2 p-0.5 text-claude-subtext/40 hover:text-claude-text transition-colors"
-                >
-                  <X size={14} />
-                </button>
-              )}
-            </div>
+          <DocumentSearchBar
+            searchQuery={searchQuery}
+            onSearchQueryChange={setSearchQuery}
+            onSearch={handleSearch}
+            onClearSearch={handleClearSearch}
+            filterType={filterType}
+            onFilterTypeChange={setFilterType}
+            viewMode={viewMode}
+            onViewModeChange={handleSetViewMode}
+          />
 
-            {/* Type filter */}
-            <select
-              value={filterType}
-              onChange={(e) => setFilterType(e.target.value as DocType | '')}
-              className="px-3 py-2.5 bg-white border border-claude-border rounded-xl text-sm text-claude-text focus:outline-none focus:border-claude-subtext/40 transition-colors font-sans"
-              title="Фільтр за типом документа"
-            >
-              <option value="">Всі типи</option>
-              {(Object.keys(DOC_TYPE_LABELS) as DocType[]).map((t) => (
-                <option key={t} value={t}>
-                  {DOC_TYPE_LABELS[t]}
-                </option>
-              ))}
-            </select>
+          <BatchActionsBar
+            selectedCount={selectedIds.size}
+            onBatchMove={handleBatchMove}
+            onBatchDelete={handleBatchDelete}
+            onClearSelection={clearSelection}
+          />
 
-            {/* View mode toggle */}
-            <div className="flex border border-claude-border rounded-xl overflow-hidden">
-              <button
-                onClick={() => handleSetViewMode('list')}
-                className={`p-2.5 transition-colors ${
-                  viewMode === 'list'
-                    ? 'bg-claude-text text-white'
-                    : 'bg-white text-claude-subtext hover:bg-claude-bg'
-                }`}
-                title="Таблиця"
-              >
-                <List size={16} />
-              </button>
-              <button
-                onClick={() => handleSetViewMode('grid')}
-                className={`p-2.5 transition-colors ${
-                  viewMode === 'grid'
-                    ? 'bg-claude-text text-white'
-                    : 'bg-white text-claude-subtext hover:bg-claude-bg'
-                }`}
-                title="Сітка"
-              >
-                <LayoutGrid size={16} />
-              </button>
-            </div>
-          </div>
-
-          {/* Batch actions bar */}
-          <AnimatePresence>
-            {selectedIds.size > 0 && (
-              <motion.div
-                initial={{ opacity: 0, y: -8 }}
-                animate={{ opacity: 1, y: 0 }}
-                exit={{ opacity: 0, y: -8 }}
-                className="mb-4 flex items-center gap-3 px-4 py-2.5 bg-claude-accent/5 border border-claude-accent/20 rounded-xl"
-              >
-                <span className="text-sm font-medium text-claude-text font-sans">
-                  Вибрано: {selectedIds.size}
-                </span>
-                <div className="flex items-center gap-2 ml-auto">
-                  <button
-                    onClick={handleBatchMove}
-                    className="inline-flex items-center gap-1.5 px-3 py-1.5 text-xs font-medium border border-claude-border rounded-lg hover:bg-claude-bg transition-colors font-sans"
-                  >
-                    <FolderInput size={12} />
-                    Перемістити
-                  </button>
-                  <button
-                    onClick={handleBatchDelete}
-                    className="inline-flex items-center gap-1.5 px-3 py-1.5 text-xs font-medium text-red-600 border border-red-200 rounded-lg hover:bg-red-50 transition-colors font-sans"
-                  >
-                    <Trash2 size={12} />
-                    Видалити
-                  </button>
-                  <button
-                    onClick={() => setSelectedIds(new Set())}
-                    className="p-1 text-claude-subtext/40 hover:text-claude-text transition-colors"
-                    title="Зняти виділення"
-                  >
-                    <X size={14} />
-                  </button>
-                </div>
-              </motion.div>
-            )}
-          </AnimatePresence>
-
-          {/* Folder Navigator */}
           {(currentFolderPath || folders.length > 0) && (
             <FolderNavigator
               currentPath={currentFolderPath}
@@ -990,132 +334,48 @@ export function DocumentsPage() {
             />
           )}
 
-          {/* Document List */}
-          {loading ? (
-            <SkeletonRows />
-          ) : documents.length === 0 ? (
-            isSearchActive ? (
-              /* Search no results */
-              <div className="text-center py-20">
-                <SearchX size={48} className="mx-auto mb-4 text-claude-subtext/20" />
-                <h3 className="text-lg font-semibold text-claude-text mb-2 font-sans">
-                  Нічого не знайдено
-                </h3>
-                <p className="text-sm text-claude-subtext/60 font-sans mb-4">
-                  За запитом «{searchQuery}» не знайдено документів
-                </p>
-                <button
-                  onClick={handleClearSearch}
-                  className="inline-flex items-center gap-2 px-4 py-2 bg-white border border-claude-border text-claude-text rounded-xl text-sm font-medium hover:bg-claude-bg transition-colors font-sans"
-                >
-                  <X size={14} />
-                  Очистити пошук
-                </button>
-              </div>
-            ) : (
-              /* Empty state */
-              <div className="text-center py-20">
-                <FileText size={48} className="mx-auto mb-4 text-claude-subtext/20" />
-                <h3 className="text-lg font-semibold text-claude-text mb-2 font-sans">
-                  Немає документів
-                </h3>
-                <p className="text-sm text-claude-subtext/60 font-sans">
-                  Завантажте документи за допомогою кнопок вище або перетягніть файли
-                </p>
-              </div>
-            )
-          ) : viewMode === 'list' ? (
-            <DocumentTable
-              documents={documents}
-              sortBy={sortBy}
-              sortOrder={sortOrder}
-              onSort={handleSort}
-              onDocumentClick={handleDocumentClick}
-              onView={handleDocumentClick}
-              onEdit={handleEditOpen}
-              onDelete={setDeleteTarget}
-              onMove={handleMoveOpen}
-              selectedIds={selectedIds}
-              onToggleSelect={handleToggleSelect}
-              onToggleAll={handleToggleAll}
-            />
-          ) : (
-            <DocumentGrid
-              documents={documents}
-              onDocumentClick={handleDocumentClick}
-              onView={handleDocumentClick}
-              onEdit={handleEditOpen}
-              onDelete={setDeleteTarget}
-              onMove={handleMoveOpen}
-              selectedIds={selectedIds}
-              onToggleSelect={handleToggleSelect}
-              sortBy={sortBy}
-              sortOrder={sortOrder}
-              onSort={handleSort}
-            />
-          )}
-
-          {/* Pagination footer — always show when there are results */}
-          {documents.length > 0 && (hasMore || hasPrev || isSearchActive) && (
-            <div className="flex items-center justify-between mt-4">
-              <span className="text-xs text-claude-subtext/50 font-sans">
-                {isSearchActive ? `Знайдено: ` : ''}{offset + 1}–{Math.min(offset + PAGE_SIZE, totalDocs)} з {totalDocs}
-              </span>
-              <div className="flex gap-2">
-                <button
-                  onClick={() => setOffset((prev) => Math.max(0, prev - PAGE_SIZE))}
-                  disabled={!hasPrev}
-                  className="px-4 py-2 rounded-xl text-sm font-sans font-medium transition-all disabled:opacity-40 disabled:cursor-not-allowed bg-white border border-claude-border text-claude-text hover:bg-claude-bg"
-                >
-                  Назад
-                </button>
-                <button
-                  onClick={() => setOffset((prev) => prev + PAGE_SIZE)}
-                  disabled={!hasMore}
-                  className="px-4 py-2 rounded-xl text-sm font-sans font-medium transition-all disabled:opacity-40 disabled:cursor-not-allowed bg-white border border-claude-border text-claude-text hover:bg-claude-bg"
-                >
-                  Далі
-                </button>
-              </div>
-            </div>
-          )}
-
-          {/* Simple count when only one page */}
-          {documents.length > 0 && !hasMore && !hasPrev && !isSearchActive && (
-            <div className="mt-4 text-center">
-              <span className="text-xs text-claude-subtext/50 font-sans">
-                {totalDocs} документів
-              </span>
-            </div>
-          )}
+          <DocumentListContainer
+            documents={documents}
+            loading={loading}
+            viewMode={viewMode}
+            sortBy={sortBy}
+            sortOrder={sortOrder}
+            onSort={handleSort}
+            onDocumentClick={preview.handleDocumentClick}
+            onEdit={handleEditOpen}
+            onDelete={setDeleteTarget}
+            onMove={handleMoveOpen}
+            selectedIds={selectedIds}
+            onToggleSelect={handleToggleSelect}
+            onToggleAll={handleToggleAll}
+            offset={offset}
+            setOffset={setOffset}
+            totalDocs={totalDocs}
+            pageSize={PAGE_SIZE}
+            searchQuery={searchQuery}
+            onClearSearch={handleClearSearch}
+          />
         </div>
       </div>
 
-      {/* Document Preview Modal */}
       <DocumentViewerModal
-        isOpen={previewOpen}
-        onClose={() => {
-          setPreviewOpen(false);
-          setPreviewDoc(null);
-          setPreviewIndex(-1);
-          setPreviewDeletePending(false);
-        }}
-        item={previewLoading ? {
+        isOpen={preview.previewOpen}
+        onClose={preview.closePreview}
+        item={preview.previewLoading ? {
           type: 'document',
           title: 'Завантаження...',
           content: '',
-        } : previewDoc}
+        } : preview.previewDoc}
         onSaveOcrText={handleSaveOcrText}
-        onPrevious={handlePreviewPrevious}
-        onNext={handlePreviewNext}
-        hasPrevious={previewIndex > 0 || offset > 0}
-        hasNext={previewIndex < documents.length - 1 || offset + PAGE_SIZE < totalDocs}
-        currentIndex={offset + previewIndex}
+        onPrevious={preview.handlePreviewPrevious}
+        onNext={preview.handlePreviewNext}
+        hasPrevious={preview.previewIndex > 0 || offset > 0}
+        hasNext={preview.previewIndex < documents.length - 1 || offset + PAGE_SIZE < totalDocs}
+        currentIndex={offset + preview.previewIndex}
         totalCount={totalDocs}
-        onDelete={handlePreviewDelete}
+        onDelete={preview.handlePreviewDelete}
       />
 
-      {/* Delete Confirmation Modal */}
       <DeleteConfirmModal
         target={deleteTarget}
         deleting={deleting}
@@ -1123,7 +383,6 @@ export function DocumentsPage() {
         onCancel={() => setDeleteTarget(null)}
       />
 
-      {/* Edit Document Modal */}
       <EditDocumentModal
         target={editTarget}
         editText={editText}
@@ -1139,7 +398,6 @@ export function DocumentsPage() {
         totalCount={documents.length}
       />
 
-      {/* Move Document Modal */}
       <MoveDocumentModal
         target={moveTarget}
         moveFolder={moveFolder}
@@ -1154,40 +412,4 @@ export function DocumentsPage() {
       />
     </div>
   );
-}
-
-// Recursively read FileSystemEntry items (for drag-and-drop folder support)
-async function readEntries(entries: any[]): Promise<File[]> {
-  const files: File[] = [];
-
-  async function processEntry(entry: any): Promise<void> {
-    if (entry.isFile) {
-      const file = await new Promise<File>((resolve) => entry.file(resolve));
-      files.push(file);
-    } else if (entry.isDirectory) {
-      const reader = entry.createReader();
-      const subEntries = await new Promise<any[]>((resolve) => {
-        const allEntries: any[] = [];
-        const readBatch = () => {
-          reader.readEntries((batch: any[]) => {
-            if (batch.length === 0) {
-              resolve(allEntries);
-            } else {
-              allEntries.push(...batch);
-              readBatch();
-            }
-          });
-        };
-        readBatch();
-      });
-      for (const sub of subEntries) {
-        await processEntry(sub);
-      }
-    }
-  }
-
-  for (const entry of entries) {
-    await processEntry(entry);
-  }
-  return files;
 }
